+++ pkg/event-sourcing/projection/performance_benchmark_test.go
@@ -138,6 +138,7 @@ func projectionPerformanceRunner(
 
 			return nil
 		},
+		Guard:     PermitReplay,
 		BatchSize: size,
 	})
 	if err != nil {